Added shard player window
Moved shard drawing logic to a common class
